Fix overflow in get_token_alloc leading to OOM
A signed integer overflow occurs in get_token_alloc() when the
VP9 encoder is initialized with large frame dimensions.
This causes incorrect buffer size calculation and OOM.

This commit fixes the overflow by using 64-bit arithmetic
in get_token_alloc() and updating callers to avoid truncation.

Added a unit test LargeDimensionsTokenAllocOverflow to verify
that large dimensions are handled without crashing or OOMing.

Bug: 505286908

+++ vp9/encoder/vp9_encoder.h
@@ -1064,7 +1064,7 @@ static INLINE YV12_BUFFER_CONFIG *get_ref_frame_buffer(
                                 : NULL;
 }
 
-static INLINE int get_token_alloc(int mb_rows, int mb_cols) {
+static INLINE int64_t get_token_alloc(int mb_rows, int mb_cols) {
   // TODO(JBB): double check we can't exceed this token count if we have a
   // 32x32 transform crossing a boundary at a multiple of 16.
   // mb_rows, cols are in units of 16 pixels. We assume 3 planes all at full
@@ -1076,12 +1076,12 @@ static INLINE int get_token_alloc(int mb_rows, int mb_cols) {
       ALIGN_POWER_OF_TWO(mb_rows, MI_BLOCK_SIZE_LOG2 - 1);
   const int aligned_mb_cols =
       ALIGN_POWER_OF_TWO(mb_cols, MI_BLOCK_SIZE_LOG2 - 1);
-  return aligned_mb_rows * aligned_mb_cols * (16 * 16 * 3 + 4);
+  return (int64_t)aligned_mb_rows * aligned_mb_cols * (16 * 16 * 3 + 4);
 }
 
 // Get the allocated token size for a tile. It does the same calculation as in
 // the frame token allocation.
-static INLINE int allocated_tokens(TileInfo tile) {
+static INLINE int64_t allocated_tokens(TileInfo tile) {
   int tile_mb_rows = (tile.mi_row_end - tile.mi_row_start + 1) >> 1;
   int tile_mb_cols = (tile.mi_col_end - tile.mi_col_start + 1) >> 1;
